Pages, Forms, and DBs
Most of the code this far is identical to what Corey Schafer writes in his Flask tutorial series, up through the Database tutorial (Part 4). I want to learn the basics of Flask first and have a template to use for the site that will hold the Mastermind game (and hopefully more projects in the future).